перенес renderCard, создал renderCards
