+++ src/boardController.js
@@ -1,4 +1,4 @@
-import {render, removeComponent, replaceCard} from './render.js';
+import {render, renderCard, removeComponent} from './render.js';
 import Sorter from './components/sorter.js';
 import NoTask from './components/no-task.js';
 import BoardTasks from './components/board-tasks.js';
@@ -6,40 +6,18 @@ import Card from './components/card.js';
 import CardEdit from './components/cardEdit.js';
 import LoadMoreButton from './components/button.js';
 import {isNotActive} from './utils.js';
-import {CARD_SHOWING, KEY_CODE_ESC} from './const.js';
+import {CARD_SHOWING} from './const.js';
 
 
-const renderCard = function (cardComponent, cardEditComponent, container) {
-  const escDownHandler = function (evt) {
-    if (evt.keyCode === KEY_CODE_ESC) {
-      evt.preventDefault();
-      replaceCard(cardComponent, cardEditComponent);
-      document.removeEventListener(`keydown`, escDownHandler);
-    }
-  };
-  const editHandler = function (evt, cardEdit, card) {
-    evt.preventDefault();
-    replaceCard(cardEdit, card);
+const renderCards = function (taskList, container, counter, cardForShowing) {
+  taskList.slice(counter, counter + cardForShowing)
+    .forEach((task) => {
+      const card = new Card(task);
+      const cardEdit = new CardEdit(task);
 
-    document.addEventListener(`keydown`, escDownHandler);
-  };
-  const submitHandler = function (evt, card, cardEdit) {
-    evt.preventDefault();
-    replaceCard(card, cardEdit);
-
-    document.removeEventListener(`keydown`, escDownHandler);
-  };
-
-  cardComponent.setEditHandler(function (evt) {
-    editHandler(evt, cardEditComponent, cardComponent);
-  });
-  cardEditComponent.setSubmitHandler(function (evt) {
-    submitHandler(evt, cardComponent, cardEditComponent);
-  });
-
-  render(cardComponent, container);
+      renderCard(card, cardEdit, container);
+    });
 };
-
 
 class BoardController {
   constructor(container) {
@@ -61,26 +39,13 @@ class BoardController {
     render(this._sortComponent, this._container);
     render(this._boardTasks, this._container);
 
-    while (this._showedTaskCount < tasks.length && this._showedTaskCount < CARD_SHOWING) {
-      this._cardComponent = new Card(tasks[this._showedTaskCount]);
-      this._cardEditComponent = new CardEdit(tasks[this._showedTaskCount]);
-
-      renderCard(this._cardComponent, this._cardEditComponent, this._boardTasks);
-      this._showedTaskCount++;
-    }
+    renderCards(tasks, this._boardTasks, this._showedTaskCount, CARD_SHOWING);
+    this._showedTaskCount += CARD_SHOWING;
 
     const loadMoreClickHandler = () => {
-      let counter = 0;
-      while (this._showedTaskCount < tasks.length && counter < CARD_SHOWING) {
-        this._cardComponent = new Card(tasks[this._showedTaskCount]);
-        this._cardEditComponent = new CardEdit(tasks[this._showedTaskCount]);
-
-        renderCard(this._cardComponent, this._cardEditComponent, this._boardTasks);
-
-        this._showedTaskCount++;
-        counter++;
-      }
-      if (this._showedTaskCount === tasks.length) {
+      renderCards(tasks, this._boardTasks, this._showedTaskCount, CARD_SHOWING);
+      this._showedTaskCount += CARD_SHOWING;
+      if (this._showedTaskCount >= tasks.length) {
         removeComponent(this._loadMoreButton);
       }
     };
